Move Parser, Selector & Merger to gtfs_static package and update imports
