+++ backend/src/scripts/visualize_stops.py
@@ -5,8 +5,8 @@ from pathlib import Path
 import matplotlib.pyplot as plt
 import networkx as nx
 
-from src.data_provider.Merger import Merger
-from src.data_provider.Parser import Parser
+from src.data_provider.gtfs_static.Merger import Merger
+from src.data_provider.gtfs_static.Parser import Parser
 
 if __name__ == '__main__':
     data_dir = Path(__file__).parent.parent / 'data_provider' / 'data'
